refactor: move components to common directory for better organization
