+++ src/app/components/common/NewRecipe/components/RecipeIngredientsContent.tsx
