+++ src/app/components/common/NewRecipe/components/RecipeStepsContent.tsx
