+++ src/app/components/common/NewRecipe/components/SearchProductModal/ExistingProduct.tsx
