+++ src/app/components/common/NewRecipe/components/SearchProductModal/ProductSearchInput.tsx
