+++ src/app/components/common/NewRecipe/components/SearchProductModal/helper.ts
