+++ src/app/components/common/NewRecipe/components/SearchProductModal/style.module.scss
