+++ src/app/components/common/NewRecipe/components/style.module.scss
