+++ src/app/components/common/NewRecipe/style.module.scss
