+++ src/app/components/common/RecipeImgAndInfoCard/index.tsx
