+++ src/app/components/common/Search/helper.ts
