+++ src/app/components/common/Search/style.module.scss
