+++ src/app/sn/search/RecipeList/index.tsx
@@ -2,7 +2,7 @@ import style from './style.module.scss';
 
 import { RecipeSimple } from '@/services/recipe/type';
 
-import RecipeImgAndInfoCard from '@/app/components/RecipeImgAndInfoCard';
+import RecipeImgAndInfoCard from '@/app/components/common/RecipeImgAndInfoCard';
 
 interface Props {
   recipes: RecipeSimple[];
